Refactor cakes, bakeries, bakers
