+++ src/app/bakeries/mod.rs
@@ -1,3 +1,2 @@
 pub mod bakery;
 pub mod handler;
-pub use super::pool;
