+++ src/app/bakers/baker.rs
@@ -1,12 +1,11 @@
 use sea_orm::entity::prelude::*;
 use rocket::serde::{Serialize, Deserialize};
-use crate::app::cakes as cakes;
-use crate::app::bakeries as bakeries;
+use crate::app::cakes::cake as cake;
+use crate::app::bakeries::bakery as bakery;
 use crate::app::cakes_bakers as cakes_bakers;
-use crate::app::lineitems as lineitems;
-use crate::app::customers as customers;
-use crate::app::orders as orders;
-
+use crate::app::lineitems::lineitem as lineitem;
+use crate::app::customers::customer as customer;
+use crate::app::orders::order as order;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
 #[serde(crate = "rocket::serde")]
@@ -22,22 +21,22 @@ pub struct Model {
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
-        belongs_to = "bakeries::bakery::Entity",
+        belongs_to = "bakery::Entity",
         from = "Column::BakeryId",
-        to = "bakeries::bakery::Column::Id",
+        to = "bakery::Column::Id",
         on_update = "Cascade",
         on_delete = "Cascade"
     )]
     Bakery,
 }
 
-impl Related<bakeries::bakery::Entity> for Entity {
+impl Related<bakery::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Bakery.def()
     }
 }
 
-impl Related<cakes::cake::Entity> for Entity {
+impl Related<cake::Entity> for Entity {
     fn to() -> RelationDef {
         cakes_bakers::Relation::Cake.def()
     }
@@ -52,15 +51,15 @@ pub struct BakedForCustomer;
 impl Linked for BakedForCustomer {
     type FromEntity = Entity;
 
-    type ToEntity = customers::customer::Entity;
+    type ToEntity = customer::Entity;
 
     fn link(&self) -> Vec<RelationDef> {
         vec![
             cakes_bakers::Relation::Baker.def().rev(),
             cakes_bakers::Relation::Cake.def(),
-            lineitems::lineitem::Relation::Cake.def().rev(),
-            lineitems::lineitem::Relation::Order.def(),
-            orders::order::Relation::Customer.def(),
+            lineitem::Relation::Cake.def().rev(),
+            lineitem::Relation::Order.def(),
+            order::Relation::Customer.def(),
         ]
     }
 }
