+++ src/app/bakers/mod.rs
@@ -1,3 +1,2 @@
 pub mod baker;
 pub mod handler;
-pub use super::pool;
