+++ src/app/cakes/handler.rs
@@ -2,10 +2,9 @@ use rocket::http::Status;
 use rocket::serde::json::{json, Json, Value};
 use rocket_db_pools::{Connection};
 use sea_orm::{entity::*, query::*};
-// use crate::app::bakery_chain::*;
-use super::cake::Entity as Cake;
 
-use super::pool;
+use super::cake::Entity as Cake;
+use crate::pool;
 
 #[get("/")]
 pub async fn all(connection: Connection<pool::Db>) -> Result<Json<Vec<super::cake::Model>>, Status> {
