+++ src/app/cakes/mod.rs
@@ -1,3 +1,2 @@
 pub mod handler;
 pub mod cake;
-pub use super::pool;
